Add Private AI Keys page and update sidebar navigation
- Create new Private AI Keys page with region-based key management
- Implement key creation dialog with region selection
- Add credential visibility toggle for database and LLM credentials
- Update sidebar layout with new navigation structure and icons
- Remove deprecated nav-main component and integrate navigation into sidebar layout
